+++ CHANGELOG.md
@@ -50,7 +50,7 @@ Ref: https://keepachangelog.com/en/1.0.0/
 * Updated documentation for each module to work with docusaurus [PR 1763](https://github.com/provenance-io/provenance/pull/1763).
 * Create a default market in `make run`, `localnet`, `devnet` and the `provenanced testnet` command [#1757](https://github.com/provenance-io/provenance/issues/1757).
 * Remove the rust upgrade handlers [PR 1774](https://github.com/provenance-io/provenance/pull/1774).
-* Add StoreLoader wrapper to check validator settings [#1792](https://github.com/provenance-io/provenance/pull/1792).
+* Add StoreLoader wrapper to check configuration settings [#1792](https://github.com/provenance-io/provenance/pull/1792).
 
 ### Dependencies